src/requirements/base.txt
Django
psycopg2-binary
celery
django-celery-results
Pillow
django-ckeditor
djangorestframework
django-rest-knox